CRT-50 rename: character.dto 패키지 이름을 character.api.dto로 변경
