--- src/main/java/knu/ridi/charactalk/character/dto/CharacterResponse.java
@@ -1,25 +1,0 @@
-package knu.ridi.charactalk.character.dto;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import knu.ridi.charactalk.character.domain.Character;
-
-@Schema(description = "캐릭터 응답 DTO")
-public record CharacterResponse(
-        @Schema(description = "캐릭터 ID", example = "1")
-        Long id,
-        @Schema(description = "캐릭터 이름", example = "신데렐라")
-        String name,
-        @Schema(description = "캐릭터 설명", example = "신데렐라는 아름다운 소녀로, 마법의 도움으로 왕자와 만난다.")
-        String description,
-        @Schema(description = "캐릭터 이미지 URL", example = "https://example.com/image.jpg")
-        String imageUrl
-) {
-    public static CharacterResponse from(Character character) {
-        return new CharacterResponse(
-                character.getId(),
-                character.getName(),
-                character.getDescription(),
-                character.getImageUrl()
-        );
-    }
-}
